lots of improvements for the updater

+++ classes/WpMatomo/Updater.php
@@ -15,12 +15,15 @@ use Piwik\Option;
 use Piwik\Plugins\Installation\ServerFilesGenerator;
 use Piwik\SettingsServer;
 use Piwik\Version;
+use WpMatomo\Updater\UpdateInProgressException;
 
 if ( ! defined( 'ABSPATH' ) ) {
 	exit; // if accessed directly
 }
 
 class Updater {
+	const LOCK_NAME = 'matomo_updater';
+
 	/**
 	 * @var Settings
 	 */
@@ -36,7 +39,7 @@ class Updater {
 		$this->logger   = new Logger();
 	}
 
-	private function load_plugin_functions() {
+	public function load_plugin_functions() {
 		if ( ! function_exists( 'get_plugin_data' ) ) {
 			require_once ABSPATH . '/wp-admin/includes/plugin.php';
 		}
@@ -44,13 +47,13 @@ class Updater {
 		return function_exists( 'get_plugin_data' );
 	}
 
-	public function update_if_needed() {
+	public function get_plugins_requiring_update()
+	{
 		if ( ! $this->load_plugin_functions() ) {
-			return;
+			return [];
 		}
 
-		$executed_updates = array();
-
+		$keys = [];
 		$plugin_files = $GLOBALS['MATOMO_PLUGIN_FILES'];
 		if ( ! in_array( MATOMO_ANALYTICS_FILE, $plugin_files, true ) ) {
 			$plugin_files[] = MATOMO_ANALYTICS_FILE;
@@ -65,29 +68,42 @@ class Updater {
 			$installed_ver = get_option( $key );
 			if ( ! $installed_ver || $installed_ver !== $plugin_data['Version'] ) {
 				if ( ! Installer::is_intalled() ) {
-					return;
+					return [];
 				}
-
-				try {
-					$this->update();
-				} catch ( \Exception $e ) {
-					$this->logger->log_exception( 'plugin_update', $e );
-					throw $e;
-				}
-				$executed_updates[] = $key;
-
-				// we're scheduling another update in case there are some dimensions to be updated or anything
-				// we do not do this in the "update" method as otherwise we might be calling this recursively...
-				// it is possible that because the plugins need to be reloaded etc that those updates are not executed right
-				// away but need an actual reload and cache clearance etc
-				wp_schedule_single_event( time() + 5, ScheduledTasks::EVENT_UPDATE );
-
-				update_option( $key, $plugin_data['Version'] );
-
-				// we make sure to delete cache even if no component was updated eg there may be translation updates etc
-				// and caches need to be invalidated
-				Filesystem::deleteAllCacheOnUpdate();
+				$keys[$key] = $plugin_data['Version'];
 			}
+		}
+
+		return $keys;
+	}
+
+	public function update_if_needed() {
+		$executed_updates = array();
+
+		$plugins_requiring_update = $this->get_plugins_requiring_update();
+		foreach ($plugins_requiring_update as $key => $plugin_version) {
+			try {
+				$this->update();
+			} catch ( UpdateInProgressException $e ) {
+				$this->logger->log( 'Matomo update is already in progress');
+				return; // we also don't execute any further update as they should be executed in another process
+			}catch ( \Exception $e ) {
+				$this->logger->log_exception( 'plugin_update', $e );
+				continue;
+			}
+			$executed_updates[] = $key;
+
+			// we're scheduling another update in case there are some dimensions to be updated or anything
+			// we do not do this in the "update" method as otherwise we might be calling this recursively...
+			// it is possible that because the plugins need to be reloaded etc that those updates are not executed right
+			// away but need an actual reload and cache clearance etc
+			wp_schedule_single_event( time() + 15, ScheduledTasks::EVENT_UPDATE );
+
+			update_option( $key, $plugin_version );
+
+			// we make sure to delete cache even if no component was updated eg there may be translation updates etc
+			// and caches need to be invalidated
+			Filesystem::deleteAllCacheOnUpdate();
 		}
 
 		return $executed_updates;
@@ -105,7 +121,7 @@ class Updater {
 			}
 
 			if ( ! empty( $plugin_data['Version'] )
-				&& ! in_array( $plugin_data['Version'], $history, true ) ) {
+			     && ! in_array( $plugin_data['Version'], $history, true ) ) {
 				// this allows us to see which versions of matomo the user was using before this update so we better understand
 				// which version maybe regressed something
 				array_unshift( $history, $plugin_data['Version'] );
@@ -125,8 +141,8 @@ class Updater {
 
 		\Piwik\Access::doAsSuperUser(
 			function () {
-					self::update_components();
-					self::update_components();
+				self::update_components();
+				self::update_components();
 			}
 		);
 
@@ -153,6 +169,37 @@ class Updater {
 		}
 	}
 
+	public function is_upgrade_in_progress() {
+		if ( ! self::load_upgrader() ) {
+			return 'no upgrader';
+		}
+
+		if (self::lock(2)) {
+			// we can get the lock meaning no update is in progress
+			self::unlock();
+			return false;
+		}
+
+		return true;
+	}
+
+	private static function load_upgrader() {
+		if (!class_exists('\WP_Upgrader', false)) {
+			@include_once ABSPATH . 'wp-admin/includes/class-wp-upgrader.php';
+		}
+		return class_exists('\WP_Upgrader', false);
+	}
+
+	public static function lock($time)
+	{
+		return self::load_upgrader() && \WP_Upgrader::create_lock(self::LOCK_NAME, $time);
+	}
+
+	public static function unlock()
+	{
+		self::load_upgrader() && \WP_Upgrader::release_lock(self::LOCK_NAME);
+	}
+
 	private static function update_components() {
 		$updater                     = new \Piwik\Updater();
 		$components_with_update_file = $updater->getComponentUpdates( );
@@ -161,13 +208,30 @@ class Updater {
 			return false;
 		}
 
+		if (!self::lock(60*4)) {
+			// prevent the upgrade from being started several times at once
+			// we lock for 4 minutes. In case of major Matomo upgrades the upgrade may take much longer but it should be
+			// safe in this case to run the upgrade several times
+			throw new UpdateInProgressException();
+		}
+
 		SettingsServer::setMaxExecutionTime(0);
 
 		if (function_exists('ignore_user_abort')) {
 			@ignore_user_abort(true);
 		}
 
-        $updater->updateComponents( $components_with_update_file );
+		try {
+			$result = $updater->updateComponents( $components_with_update_file );
+		} catch (\Exception $e) {
+			self::unlock();
+			throw $e;
+		}
+		self::unlock();
+
+		if (!empty($result['errors'])) {
+			throw new \Exception('Error while updating components: ' . implode(', ' , $result['errors']));
+		}
 
 		\Piwik\Updater::recordComponentSuccessfullyUpdated( 'core', Version::VERSION );
 		Filesystem::deleteAllCacheOnUpdate();
